+++ pwsh/dev/devAzGovVizParallel.ps1
@@ -277,10 +277,10 @@ Param
     $Product = 'AzGovViz',
 
     [string]
-    $AzAPICallVersion = '1.1.13',
+    $AzAPICallVersion = '1.1.15',
 
     [string]
-    $ProductVersion = 'v6_major_20220603_1',
+    $ProductVersion = 'v6_major_20220610_2',
 
     [string]
     $GithubRepository = 'aka.ms/AzGovViz',
@@ -562,11 +562,12 @@ $null = $modules.Add([PSCustomObject]@{
 
 if ($DoPSRule) {
     
-    #temporary workaround / PSRule/Azure DevOps Az.Resources module requirements
+    <#temporary workaround / PSRule/Azure DevOps Az.Resources module requirements
     if ($env:SYSTEM_TEAMPROJECTID -and $env:BUILD_REPOSITORY_ID) {
         $PSRuleVersion = '1.14.3'
         Write-Host "Running in Azure DevOps; enforce PSRule version '$PSRuleVersion' (Az.Resources dependency on latest PSRule)"
     }
+    #>
 
     $null = $modules.Add([PSCustomObject]@{
             ModuleName         = 'PSRule.Rules.Azure'
@@ -714,7 +715,7 @@ if ($azAPICallConf['htParameters'].HierarchyMapOnly -eq $false) {
     $htServicePrincipals = [System.Collections.Hashtable]::Synchronized((New-Object System.Collections.Hashtable)) #@{}
     $htDailySummary = @{}
     $arrayDefenderPlans = [System.Collections.ArrayList]::Synchronized((New-Object System.Collections.ArrayList))
-    $arrayDefenderPlansSubscriptionNotRegistered = [System.Collections.ArrayList]::Synchronized((New-Object System.Collections.ArrayList))
+    $arrayDefenderPlansSubscriptionsSkipped = [System.Collections.ArrayList]::Synchronized((New-Object System.Collections.ArrayList))
     $arrayUserAssignedIdentities4Resources = [System.Collections.ArrayList]::Synchronized((New-Object System.Collections.ArrayList))
     $htSubscriptionsRoleAssignmentLimit = [System.Collections.Hashtable]::Synchronized((New-Object System.Collections.Hashtable)) #@{}
     if ($azAPICallConf['htParameters'].NoMDfCSecureScore -eq $false) {